Only show single image and force images to top before flickity is initialised
When the page loads (before flickity is initialised) all listItems are rendered below each other and push down other elements,
when flickity is initialised all listItems are positioned next to each other and the other elements jumps back up.
To prevent this "jumping-behaviour" all the listItems are now immediately positioned to the top (absolute) and have visibility set to hidden using css.
Only the first listItem will not be positioned absolute and will be visible.
Simply setting display to none on all but the first listItem does not work since flickity needs to be able to get the dimensions of the element when initialising.

Styling for listItem will now also work when no uniquePrefix is set by using .kwcBem__listItem in css.